+++ test/conftest.py
@@ -13,7 +13,7 @@ import pytest
 
 # noinspection PyTrailingSemicolon
 import matplotlib; matplotlib.use("Agg")  # noqa, yapf: disable
-from justice import mmap_array, supernova_data
+from justice.datasets import supernova_data, mmap_array
 
 _testdata_dir = os.path.abspath(os.path.join(os.path.abspath(__file__), "../testdata"))
